requirements.txt
# Core ML packages
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.7.1
matplotlib==3.10.6
seaborn==0.13.2
joblib==1.5.2

# API
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7

# Development
jupyter==1.1.1
ipykernel==6.30.1
python-dotenv==1.1.1
requests==2.32.5

# Database
pymongo==4.14.1
psycopg2-binary==2.9.10

# Scheduling
APScheduler==3.11.0

# ML MODEL
lightgbm==4.6.0

#CloudSQL Proxy 관련
cloud-sql-python-connector==1.18.4
google-auth==2.40.3
google-cloud-core==2.4.3
pg8000==1.31.5

sqlalchemy==2.0.39